requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
lmdeploy==0.8.0
openai>=1.0.0
accelerate>=1.6.0
einops>=0.8.1
fire==0.7.0
Pillow>=10.0.0
numpy>=1.21.0
scikit-learn>=1.3.0
pandas>=1.5.0
tqdm>=4.67.0
easydict==1.13